feat: implement production workflow
